contracts: add attribution (#2)

+++ test/Constants.t.sol
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: BUSL-1.1
 pragma solidity 0.8.24;
 
-uint256 constant TEST_TOKEN_MAX_SUPPLY = type(uint256).max;
-uint256 constant TEST_TOKEN_PRICE = 0.05 ether;
+uint128 constant TEST_TOKEN_MAX_SUPPLY = type(uint64).max;
+uint128 constant TEST_TOKEN_PRICE = 0.05 ether;
 string constant TEST_TOKEN_URI = "https://token.com/uri";
